no clue what I'm doing
Izzy, if you see this before I get a chance to have another go, can you ask Stephen for help?
And Stephen if you see this, as I said in my message, I don't really get how to connect the filter onclick stuff to the functions because I'm not really sure how to define the functions to begin with.